+++ provider/pkg/provider/container.go
@@ -17,7 +17,10 @@ package provider
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
+
+	"github.com/imdario/mergo"
 
 	"github.com/pulumi/pulumi/pkg/v3/codegen"
 
@@ -30,8 +33,335 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/logging"
 )
 
+// Container cluster support
+type clusterMutations struct {
+	updateHandlers map[string]containerCustomUpdateHandlerFunc
+}
+
+func (c clusterMutations) mutablePropertyPaths() []string {
+	return codegen.SortedKeys(c.updateHandlers)
+}
+
+func (c clusterMutations) update(
+	providerInstance *googleCloudProvider,
+	urn resource.URN,
+	label string,
+	res *resources.CloudAPIResource,
+	inputs resource.PropertyMap,
+	oldState resource.PropertyMap,
+) (appliedInputs resource.PropertyMap, resp map[string]interface{}, err error) {
+	// Extract old inputs from the `__inputs` field of the old state.
+	oldInputs := parseCheckpointObject(oldState)
+	// Delete the auto-injected __autonamed tag in old and new inputs to avoid spurious diffs.
+	if v, ok := oldInputs[autonamed]; ok && v.IsBool() {
+		delete(oldInputs, autonamed)
+	}
+
+	if v, ok := inputs[autonamed]; ok && v.IsBool() {
+		delete(inputs, autonamed)
+	}
+
+	diff := oldInputs.Diff(inputs)
+	batchedInputs := resource.NewObjectProperty(oldInputs)
+
+	if diff == nil {
+		logging.V(9).Infof("[%s] no diff found in update!", label)
+	} else {
+		// Calculate the detailed diff object containing information about replacements.
+		detailedDiff := calculateDetailedDiff(res, providerInstance.resourceMap.Types, diff)
+		logging.V(9).Infof("[%s] detailedDiff: %+v", label, detailedDiff)
+		logging.V(9).Infof("[%s] updateCluster oldState: %+v", label, oldState)
+
+		for _, key := range clusterUpdateOrder {
+			updateHandler, ok := c.updateHandlers[key]
+			if !ok {
+				logging.V(9).Infof("[%s] No handler found for field: %q", label, key)
+				// TODO This should be an error once all the targetted mutations are in place!
+				continue
+			}
+			logging.V(9).Infof("[%s] Invoking update for field: %q", label, key)
+			for k := range detailedDiff {
+				logging.V(9).Infof("[%s] processing diff on field: %q", label, k)
+				var diffPropPath, propPath resource.PropertyPath
+				diffPropPath, err = resource.ParsePropertyPath(k)
+				if err != nil {
+					err = fmt.Errorf("failed to parse property path: %q: %w", k, err)
+					break
+				}
+				propPath, err = resource.ParsePropertyPath(key)
+				if err != nil {
+					err = fmt.Errorf("failed to parse property path: %q: %w", k, err)
+					break
+				}
+				if propPath.Contains(diffPropPath) {
+					_ = providerInstance.host.LogStatus(context.Background(), diag.Info, urn,
+						fmt.Sprintf("Performing update for field: %q", key))
+					logging.V(9).Infof("[%s] processing diff for %q with handler for %q",
+						label, diffPropPath.String(), propPath.String())
+					logging.V(9).Infof("[%s] waiting for cluster to reach resting state", label)
+					_, err = waitForClusterRestingState(providerInstance, urn, res, inputs, oldState)
+					if err != nil {
+						break
+					}
+					newVal, _ := diffPropPath.Get(resource.NewObjectProperty(inputs))
+					logging.V(9).Infof("[%s] calling update handler", label)
+					err = updateHandler(providerInstance, urn, label, res, inputs, oldState)
+					if err != nil {
+						logging.V(9).Infof("[%s] failure updating: %+v", label, err)
+						break
+					}
+					batchedInputs, ok = diffPropPath.Add(batchedInputs, newVal)
+					if !ok {
+						err = fmt.Errorf("failed to record update at field: %q", k)
+						break
+					}
+				}
+			}
+			if err != nil {
+				break
+			}
+		}
+	}
+	appliedInputs = batchedInputs.ObjectValue()
+	uri, readErr := res.Read.Endpoint.URI(inputs.Mappable(), oldState.Mappable())
+	if readErr != nil {
+		return nil, nil, readErr
+	}
+	resp, err2 := readContainerResourceStatus(providerInstance, uri)
+	if err2 != nil {
+		logging.V(9).Infof("[%s] Failed to read cluster status: %v", label, err2)
+		return nil, nil, fmt.Errorf("failed to retrieve cluster status: %v. Previous error: %w", err2, err)
+	}
+
+	return appliedInputs, resp, err
+}
+
+func updateClusterMapping(
+	diffPropPath resource.PropertyPath,
+	apiFieldName string,
+	defaultIfMissing propValueIfMissingFunc,
+	operationSuffix string,
+	httpMethod string,
+	additionalMetadata map[string]resources.CloudAPIProperty,
+) containerCustomUpdateHandlerFunc {
+	return func(p *googleCloudProvider,
+		urn resource.URN,
+		label string,
+		res *resources.CloudAPIResource,
+		newInputs,
+		oldState resource.PropertyMap) error {
+		tok := urn.Type()
+		version := strings.Split(tok.Module().String(), "/")[1]
+		endpoint := resources.CloudAPIEndpoint{
+			Template: fmt.Sprintf(`https://container.googleapis.com/%s/`+
+				`projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}%s`, version, operationSuffix),
+			Values: append(res.Create.Endpoint.Values,
+				resources.CloudAPIResourceParam{
+					Name:    "clustersId",
+					SdkName: "name",
+					Kind:    "path",
+				}),
+		}
+		uri, err := endpoint.URI(newInputs.Mappable(), oldState.Mappable())
+		if err != nil {
+			return err
+		}
+
+		// If key is omitted, use default.
+		_, ok := diffPropPath.Get(resource.NewObjectProperty(newInputs))
+		if !ok {
+			updated, _ := diffPropPath.Add(resource.NewObjectProperty(newInputs), defaultIfMissing(oldState, diffPropPath))
+			newInputs = updated.ObjectValue()
+		}
+
+		properties := map[string]resources.CloudAPIProperty{}
+		err = mergo.Merge(&properties, additionalMetadata)
+		if err != nil {
+			return fmt.Errorf("failed to initialize metadata: %w", err)
+		}
+		properties[apiFieldName] = resources.CloudAPIProperty{SdkName: diffPropPath.String()}
+		body := p.prepareAPIInputs(newInputs, oldState, properties)
+		op, err := retryRequest(p.client, httpMethod, uri, "", body)
+		if err != nil {
+			return fmt.Errorf("error sending request: %s: %q %+v", err, uri, body)
+		}
+
+		_, err = p.waitForResourceOpCompletion(urn, res.Update.CloudAPIOperation, op)
+		if err != nil {
+			return errors.Wrapf(err, "waiting for completion")
+		}
+		return nil
+	}
+}
+
+func updateClusterNestedField(diffPropPath, targetPropPath resource.PropertyPath,
+	defaultIfMissing propValueIfMissingFunc, operationSuffix, httpMethod string) containerCustomUpdateHandlerFunc {
+	return func(p *googleCloudProvider,
+		urn resource.URN,
+		label string,
+		res *resources.CloudAPIResource,
+		newInputs,
+		oldState resource.PropertyMap) error {
+		logging.V(9).Infof("[%s] updateClusterNestedField called for property path: %q with input: %+v", label,
+			diffPropPath.String(), newInputs)
+		tok := urn.Type()
+		version := strings.Split(tok.Module().String(), "/")[1]
+		endpoint := resources.CloudAPIEndpoint{
+			Template: fmt.Sprintf(`https://container.googleapis.com/%s/`+
+				`projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}%s`, version, operationSuffix),
+			Values: append(res.Create.Endpoint.Values,
+				resources.CloudAPIResourceParam{
+					Name:    "clustersId",
+					SdkName: "name",
+					Kind:    "path",
+				}),
+		}
+		uri, err := endpoint.URI(newInputs.Mappable(), oldState.Mappable())
+		if err != nil {
+			return err
+		}
+		// Get the new config object to copy the value for fieldName from
+		newConfigField, ok := diffPropPath.Get(resource.NewObjectProperty(newInputs))
+		if !ok {
+			logging.V(9).Infof("[%s] no update found at path: %q. Assuming the field is deleted.", label,
+				diffPropPath.String())
+			newConfigField = defaultIfMissing(oldState, diffPropPath)
+		}
+		// Overwrite the config object where the one value for fieldName has been updated.
+		newConfig := resource.NewObjectProperty(resource.NewPropertyMapFromMap(nil))
+		if targetPropPath == nil {
+			targetPropPath = diffPropPath
+		}
+		// Now updatedConfigField essentially is a config object with just one field - `fieldName`
+		updatedConfigField, ok := targetPropPath.Add(newConfig, newConfigField)
+		if !ok {
+			return fmt.Errorf("failed to inject config at property path: %q", targetPropPath.String())
+		}
+		logging.V(9).Infof("[%s] staged property value: %+v", label, updatedConfigField.ObjectValue().Mappable())
+		body := p.prepareAPIInputs(updatedConfigField.ObjectValue(), oldState, map[string]resources.CloudAPIProperty{
+			"update": {},
+		})
+		logging.V(9).Infof("[%s] cluster update request body: %v", label, body)
+		op, err := retryRequest(p.client, "PUT", uri, "", body)
+		if err != nil {
+			return fmt.Errorf("error sending request: %s: %q %+v", err, uri, body)
+		}
+
+		_, err = p.waitForResourceOpCompletion(urn, res.Update.CloudAPIOperation, op)
+		if err != nil {
+			return errors.Wrapf(err, "waiting for completion")
+		}
+		return nil
+	}
+}
+
+// The ordering here loosely follows what is in the terraform provider which is itself somewhat arbitrary.
+// There is some underlying value since there are a lot overlapping dependencies on mutations (e.g. setting
+// network policies require the network policy add on etc.). This may need to change over time.
+var clusterUpdateOrder = []string{
+	"masterAuthorizedNetworksConfig",          //
+	"addonsConfig",                            //
+	"autoscaling",                             //
+	"binaryAuthorization",                     //
+	"shieldedNodes",                           //
+	"releaseChannel",                          //
+	"networkConfig.enableIntraNodeVisibility", //
+	"networkConfig.privateIpv6GoogleAccess",   //
+	"networkConfig.enableL4ilbSubsetting",     //
+	"networkConfig.defaultSnatStatus",         //
+	"maintenancePolicy",                       //
+	"locations",                               //
+	"legacyAbac.enabled",                      //
+	"monitoringService",                       //
+	"loggingService",                          //
+	"networkPolicy",                           //
+	//"nodePools",                               // TODO: Don't believe we can allow this to be changed?
+	"initialClusterVersion", //
+	// nodeVersion, // nodePools[*].version
+	"nodeConfig.imageType",   //
+	"notificationConfig",     //
+	"verticalPodAutoscaling", //
+	"meshCertificates",       //
+	"databaseEncryption",     //
+	"workloadIdentifyConfig", //
+	"identityServiceConfig",  //
+	// masterAuth has a pretty imperative API for mutations.
+	// Its mostly deprecated anyway so lets mark it replace-on-change.
+	// "masterAuth",
+	// TODO: Enable these later. See note in overrides.go.
+	//"loggingConfig",             //
+	//"monitoringConfig",          //
+	"resourceLabels",            //
+	"resourceUsageExportConfig", //
+	// Only beta
+	"podSecurityPolicyConfig", //
+	"clusterTelemetry",        //
+	// removeDefaultNodepool,
+	// Additional mutations:
+	"authenticatorGroupsConfig",              //
+	"networkConfig.datapathProvider",         //
+	"networkConfig.dnsConfig",                //
+	"nodeConfig.gcfsConfig",                  //
+	"privateClusterConfig",                   //
+	"networkConfig.serviceExternalIpsConfig", //
+}
+
+func waitForClusterRestingState(
+	p *googleCloudProvider,
+	urn resource.URN,
+	res *resources.CloudAPIResource,
+	inputs,
+	oldState resource.PropertyMap,
+) (map[string]interface{}, error) {
+	ctx, cancelFunc := context.WithTimeout(context.Background(), 15*time.Minute)
+	defer cancelFunc()
+	retryPolicy := backoff.Backoff{
+		Min:    1 * time.Second,
+		Max:    15 * time.Second,
+		Factor: 1.5,
+		Jitter: true,
+	}
+
+	uri, err := res.Read.Endpoint.URI(inputs.Mappable(), oldState.Mappable())
+	if err != nil {
+		return nil, err
+	}
+
+	for {
+		resp, err := readContainerResourceStatus(p, uri)
+		if err != nil {
+			return nil, err
+		}
+
+		status, hasStatus := resp["status"].(string)
+		if hasStatus && isClusterInRestingStatus(status) {
+			return resp, nil
+		} else {
+			_ = p.host.LogStatus(context.Background(), diag.Info, urn,
+				fmt.Sprintf("Waiting for cluster to reach resting state. "+
+					"Current status: %q", status))
+		}
+		select {
+		case <-time.After(retryPolicy.Duration()):
+		case <-ctx.Done():
+			return nil, fmt.Errorf("timeout polling for nodepool resting state: %q", uri)
+		}
+	}
+}
+
+func isClusterInRestingStatus(status string) bool {
+	switch status {
+	case "RUNNING", "DEGRADED", "ERROR":
+		return true
+	}
+	return false
+}
+
+// == Container Nodepool support ==
+
+// nodepoolMutations models the custom mutations supported for Nodepool resource.
 type nodepoolMutations struct {
-	updateHandlers map[string]nodepoolUpdateHandlerFunc
+	updateHandlers map[string]containerCustomUpdateHandlerFunc
 }
 
 func (n nodepoolMutations) mutablePropertyPaths() []string {
@@ -70,7 +400,7 @@ func (n nodepoolMutations) update(
 		logging.V(9).Infof("[%s] detailedDiff: %+v", label, detailedDiff)
 		logging.V(9).Infof("[%s] updateNodepool oldState: %+v", label, oldState)
 
-		for _, key := range updateOrder {
+		for _, key := range nodepoolUpdateOrder {
 			updateHandler, ok := n.updateHandlers[key]
 			if !ok {
 				logging.V(9).Infof("[%s] No handler found for field: %q", label, key)
@@ -125,7 +455,7 @@ func (n nodepoolMutations) update(
 	if readErr != nil {
 		return nil, nil, readErr
 	}
-	resp, err2 := readNodepoolStatus(providerInstance, uri)
+	resp, err2 := readContainerResourceStatus(providerInstance, uri)
 	if err2 != nil {
 		logging.V(9).Infof("[%s] Failed to read nodepool status: %v", label, err2)
 		return nil, nil, fmt.Errorf("failed to retrieve nodepool status: %v. Previous error: %w", err2, err)
@@ -134,7 +464,7 @@ func (n nodepoolMutations) update(
 	return appliedInputs, resp, err
 }
 
-type nodepoolUpdateHandlerFunc func(
+type containerCustomUpdateHandlerFunc func(
 	p *googleCloudProvider,
 	urn resource.URN,
 	label string,
@@ -166,7 +496,7 @@ func defaultValue(defVal resource.PropertyValue) propValueIfMissingFunc {
 
 // Following an ordering here that the terraform provider performs updates in.
 // We have a lot more mutations defined than the terraform providers does though.
-var updateOrder = []string{
+var nodepoolUpdateOrder = []string{
 	"autoscaling",
 	"config.imageType",
 	"config.workloadMetadataConfig",
@@ -191,7 +521,7 @@ func mustParsePropertyPath(pattern string) resource.PropertyPath {
 	return pp
 }
 
-func readNodepoolStatus(
+func readContainerResourceStatus(
 	p *googleCloudProvider,
 	uri string,
 ) (map[string]interface{}, error) {
@@ -225,7 +555,7 @@ func waitForNodepoolRestingState(
 	}
 
 	for {
-		resp, err := readNodepoolStatus(p, uri)
+		resp, err := readContainerResourceStatus(p, uri)
 		if err != nil {
 			return nil, err
 		}
@@ -261,16 +591,19 @@ func updateNodePoolMapping(
 	defaultIfMissing propValueIfMissingFunc,
 	operationSuffix string,
 	httpMethod string,
-) nodepoolUpdateHandlerFunc {
+) containerCustomUpdateHandlerFunc {
 	return func(p *googleCloudProvider,
 		urn resource.URN,
 		label string,
 		res *resources.CloudAPIResource,
 		newInputs,
 		oldState resource.PropertyMap) error {
+		tok := urn.Type()
+		version := strings.Split(tok.Module().String(), "/")[1]
 		endpoint := resources.CloudAPIEndpoint{
-			Template: fmt.Sprintf(`https://container.googleapis.com/v1/`+
-				`projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}/nodePools/{nodePoolsId}%s`, operationSuffix),
+			Template: fmt.Sprintf(`https://container.googleapis.com/%s/`+
+				`projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}/nodePools/{nodePoolsId}%s`,
+				version, operationSuffix),
 			Values: append(res.Create.Endpoint.Values,
 				resources.CloudAPIResourceParam{
 					Name:    "nodePoolsId",
@@ -307,7 +640,7 @@ func updateNodePoolMapping(
 }
 
 func updateNodePoolConfig(diffPropPath, targetPropPath resource.PropertyPath,
-	defaultIfMissing propValueIfMissingFunc) nodepoolUpdateHandlerFunc {
+	defaultIfMissing propValueIfMissingFunc) containerCustomUpdateHandlerFunc {
 	return func(p *googleCloudProvider,
 		urn resource.URN,
 		label string,
@@ -315,8 +648,11 @@ func updateNodePoolConfig(diffPropPath, targetPropPath resource.PropertyPath,
 		newInputs,
 		oldState resource.PropertyMap) error {
 		logging.V(9).Infof("[%s] updateNodePoolConfig called for property path: %q", label, diffPropPath.String())
+		tok := urn.Type()
+		version := strings.Split(tok.Module().String(), "/")[1]
 		endpoint := resources.CloudAPIEndpoint{
-			Template: `https://container.googleapis.com/v1/projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}/nodePools/{nodePoolsId}`,
+			Template: fmt.Sprintf(`https://container.googleapis.com/%s/`+
+				`projects/{projectsId}/locations/{locationsId}/clusters/{clustersId}/nodePools/{nodePoolsId}`, version),
 			Values: append(res.Create.Endpoint.Values,
 				resources.CloudAPIResourceParam{
 					Name:    "nodePoolsId",